requirements.txt
# API
requests>=2.31.0
python-dotenv>=1.0.0

# Google Sheets + Drive (service account)
gspread>=6.0.0
google-auth>=2.27.0
google-api-python-client>=2.111.0

# Web server (for Railway / cron endpoint)
flask>=3.0.0
gunicorn>=21.0.0
